move 'test_projection_pushdown' to runtime-datafusion
