+++ crates/datafusion-optimizer-rules/src/physical_plan/cluster/union_projection_pushdown.rs
@@ -192,121 +192,3 @@ impl PhysicalOptimizerRule for UnionProjectionPushdownOptimizer {
         true
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use arrow::datatypes::{DataType, Field, Schema};
-    use chrono::DateTime;
-    use datafusion::datasource::physical_plan::ArrowSource;
-    use datafusion::execution::object_store::ObjectStoreUrl;
-    use datafusion::physical_expr::expressions::col;
-    use datafusion::physical_optimizer::optimizer::PhysicalOptimizer;
-    use datafusion_datasource::file_groups::FileGroup;
-    use datafusion_datasource::file_scan_config::FileScanConfigBuilder;
-    use datafusion_datasource::{FileRange, PartitionedFile};
-    use object_store::{ObjectMeta, path::Path};
-
-    use std::sync::LazyLock;
-
-    #[must_use]
-    pub fn create_partitioned_file(
-        path: &str,
-        size: u64,
-        range: Option<FileRange>,
-    ) -> PartitionedFile {
-        PartitionedFile {
-            object_meta: ObjectMeta {
-                location: Path::from(path),
-                last_modified: DateTime::default(),
-                size,
-                e_tag: None,
-                version: None,
-            },
-            partition_values: vec![],
-            range,
-            statistics: None,
-            extensions: None,
-            metadata_size_hint: None,
-        }
-    }
-    fn file_scan_config_builder() -> FileScanConfigBuilder {
-        let schema = Arc::new(Schema::new(vec![
-            Field::new("id", DataType::Int64, false),
-            Field::new("name", DataType::Utf8, false),
-        ]));
-
-        FileScanConfigBuilder::new(
-            ObjectStoreUrl::parse("file://tmp/").expect("Must parse dummy URL"),
-            schema,
-            Arc::new(ArrowSource::default()),
-        )
-    }
-
-    #[must_use]
-    pub fn create_data_source_exec(files: Vec<PartitionedFile>) -> Arc<dyn ExecutionPlan> {
-        let fsc = file_scan_config_builder()
-            .with_file_group(FileGroup::new(files))
-            .build();
-
-        DataSourceExec::from_data_source(fsc)
-    }
-
-    static OPTIMIZER: LazyLock<PhysicalOptimizer> = LazyLock::new(|| {
-        PhysicalOptimizer::with_rules(vec![UnionProjectionPushdownOptimizer::new()])
-    });
-
-    fn optimize(plan: &Arc<dyn ExecutionPlan>) -> Arc<dyn ExecutionPlan> {
-        OPTIMIZER.rules.iter().fold(Arc::clone(plan), |acc, rule| {
-            rule.optimize(acc, &ConfigOptions::default())
-                .expect("Must optimize")
-        })
-    }
-
-    #[ignore = "See #8313"]
-    #[test]
-    fn test_projection_pushdown() {
-        let files = vec![
-            create_partitioned_file("file:///file4.parquet", 256_000_000, None),
-            create_partitioned_file("file:///file5.parquet", 256_000_000, None),
-        ];
-
-        let data_source_exec = create_data_source_exec(files);
-        let projection_exec = ProjectionExec::try_new(
-            vec![(
-                col("id", data_source_exec.schema().as_ref()).expect("Must bind expr"),
-                "foo".to_string(),
-            )],
-            data_source_exec,
-        )
-        .expect("Must make projection_exec");
-        let plan: Arc<dyn ExecutionPlan> = Arc::new(projection_exec);
-
-        // We start with 1 projection
-        assert_eq!(
-            SearchVisitor::collect_concrete_down::<ProjectionExec>(&plan)
-                .expect("Must collect")
-                .len(),
-            1
-        );
-
-        let optimized = optimize(&plan);
-
-        let data_source_exec_leaves =
-            SearchVisitor::collect_concrete_down::<DataSourceExec>(&optimized)
-                .expect("Must collect")
-                .len();
-
-        // Make sure we have enough leaves to test with
-        assert!(data_source_exec_leaves > 1);
-
-        // The number of projections must match the number of DataSourceExec leaves
-        assert_eq!(
-            SearchVisitor::collect_concrete_down::<ProjectionExec>(&optimized)
-                .expect("Must collect")
-                .len(),
-            data_source_exec_leaves
-        );
-    }
-}
